renamed music manager to music conductor
It does express better my intent
