+++ src/ts/cosmosJourneyer.ts
@@ -59,9 +59,9 @@ import { promptModalBoolean, alertModal, promptModalString } from "./utils/dialo
 import { FuelScoopTutorial } from "./tutorials/fuelScoopTutorial";
 import { EncyclopaediaGalacticaManager } from "./society/encyclopaediaGalacticaManager";
 import { EncyclopaediaGalacticaLocal } from "./society/encyclopaediaGalacticaLocal";
+import { MusicManager } from "./audio/musicConductor";
 import { StarSystemDatabase } from "./starSystem/starSystemDatabase";
 import { registerCustomSystems } from "./starSystem/customSystems/registerCustomSystems";
-import { MusicManager } from "./audio/musicManager";
 
 const enum EngineState {
     UNINITIALIZED,
@@ -87,7 +87,7 @@ export class CosmosJourneyer {
     readonly starSystemView: StarSystemView;
     readonly starMap: StarMap;
 
-    readonly musicManager: MusicManager;
+    readonly musicConductor: MusicManager;
 
     readonly mainMenu: MainMenu;
     readonly pauseMenu: PauseMenu;
@@ -299,7 +299,7 @@ export class CosmosJourneyer {
                 );
         });
 
-        this.musicManager = new MusicManager(this.starSystemView);
+        this.musicConductor = new MusicManager(this.starSystemView);
 
         window.addEventListener("blur", () => {
             if (!this.mainMenu?.isVisible() && !this.starSystemView.isLoadingSystem()) this.pause();
@@ -460,7 +460,7 @@ export class CosmosJourneyer {
             updateNotifications(deltaSeconds);
             AudioManager.Update(deltaSeconds);
             Sounds.Update();
-            this.musicManager.update(
+            this.musicConductor.update(
                 this.isPaused(),
                 this.activeView === this.starSystemView,
                 this.mainMenu.isVisible(),
